test: reorg encoded_len tests
Split the contents of encoded_len file into submodule
files for limit and overfow tests.
