+++ tests/src/encoded_len.rs
@@ -2,129 +2,5 @@ mod proto {
     include!(concat!(env!("OUT_DIR"), "/encoded_len.rs"));
 }
 
-use prost::alloc::vec;
-
-#[cfg(target_pointer_width = "64")]
-fn verify_overflowing_encoded_len(actual: usize, expected: u64) -> bool {
-    actual as u64 == expected
-}
-
-#[cfg(target_pointer_width = "32")]
-fn verify_overflowing_encoded_len(actual: usize, _expected: u64) -> bool {
-    // Tests calling this function are expected to panic on 32-bit platforms
-    // before this check is called. Returning true here allows the
-    // #[should_panic] tests to fail.
-    eprintln!("expected panic, but the function returned {actual}");
-    true
-}
-
-mod bool {
-    use super::*;
-    use prost::encoding::{bool, MAX_TAG};
-
-    #[test]
-    #[ignore = "allocates and fills about 666 MiB"]
-    #[cfg_attr(target_pointer_width = "32", should_panic)]
-    fn encoded_len_repeated_overflows_u32() {
-        let filler = false;
-        let filler_len = bool::encoded_len(MAX_TAG, &filler);
-        let bomb32 = vec![filler; u32::MAX as usize / filler_len + 1];
-        let encoded_len = bool::encoded_len_repeated(MAX_TAG, &bomb32);
-        let expected_len = bomb32.len() as u64 * filler_len as u64;
-        assert!(verify_overflowing_encoded_len(encoded_len, expected_len));
-    }
-}
-
-/// The tests in this module each allocate and fill more than 1 GB in memory,
-/// so they should better be run in isolation and in a sufficiently
-/// memory-budgeted environment.
-mod int32 {
-    use super::*;
-    use prost::encoding::{encoded_len_varint, int32, key_len, MAX_TAG};
-
-    #[test]
-    #[ignore = "allocates and fills more than 1 GiB"]
-    #[cfg_attr(target_pointer_width = "32", should_panic)]
-    fn encoded_len_repeated_overflows_u32() {
-        let filler = -1i32;
-        let filler_len = int32::encoded_len(MAX_TAG, &filler);
-        let bomb32 = vec![filler; u32::MAX as usize / filler_len + 1];
-        let encoded_len = int32::encoded_len_repeated(MAX_TAG, &bomb32);
-        let expected_len = bomb32.len() as u64 * filler_len as u64;
-        assert!(verify_overflowing_encoded_len(encoded_len, expected_len));
-    }
-
-    #[test]
-    #[ignore = "allocates and fills about 1.6 GiB"]
-    #[cfg_attr(target_pointer_width = "32", should_panic)]
-    fn encoded_len_packed_overflows_u32() {
-        let filler = -1i32;
-        let filler_len = encoded_len_varint(filler as u64);
-        let bomb_len = (u32::MAX as usize - key_len(MAX_TAG) - 5) / filler_len + 1;
-        let bomb32 = vec![filler; bomb_len];
-        let encoded_len = int32::encoded_len_packed(MAX_TAG, &bomb32);
-        let expected_data_len = bomb_len as u64 * filler_len as u64;
-        let expected_len = key_len(MAX_TAG) as u64
-            + encoded_len_varint(expected_data_len) as u64
-            + expected_data_len;
-        assert!(verify_overflowing_encoded_len(encoded_len, expected_len));
-    }
-}
-
-mod message {
-    use super::*;
-    use prost::encoding::{message, MAX_TAG};
-    use prost::Message;
-
-    #[derive(Clone, PartialEq, Message)]
-    struct Empty {}
-
-    #[test]
-    #[cfg_attr(target_pointer_width = "32", should_panic)]
-    fn encoded_len_repeated_overflows_u32() {
-        let filler = Empty {};
-        let filler_len = message::encoded_len(MAX_TAG, &filler);
-        let bomb32 = vec![filler; u32::MAX as usize / filler_len + 1];
-        let encoded_len = message::encoded_len_repeated(MAX_TAG, &bomb32);
-        let expected_len = bomb32.len() as u64 * filler_len as u64;
-        assert!(verify_overflowing_encoded_len(encoded_len, expected_len));
-    }
-}
-
-mod limit {
-    // Test that the limits returned by field::scalar::Ty::encoded_len_limit()
-    // in prost-derive can be reached with actual values.
-
-    macro_rules! test_limit_is_reachable {
-        ($name:ident, $value:expr, $limit:expr) => {
-            mod $name {
-                use super::super::proto;
-                use prost::Message;
-
-                #[test]
-                fn encoded_len_limit_is_reachable() {
-                    let msg = proto::Numerics {
-                        $name: $value,
-                        ..Default::default()
-                    };
-                    assert_eq!(msg.encoded_len() - 1, $limit);
-                }
-            }
-        };
-    }
-
-    test_limit_is_reachable!(int32, -1, 10);
-    test_limit_is_reachable!(int64, -1, 10);
-    test_limit_is_reachable!(uint32, u32::MAX, 5);
-    test_limit_is_reachable!(uint64, u64::MAX, 10);
-    test_limit_is_reachable!(sint32, i32::MIN, 5);
-    test_limit_is_reachable!(sint64, i64::MIN, 10);
-    test_limit_is_reachable!(fixed32, 1, 4);
-    test_limit_is_reachable!(fixed64, 1, 8);
-    test_limit_is_reachable!(sfixed32, -1, 4);
-    test_limit_is_reachable!(sfixed64, -1, 8);
-    test_limit_is_reachable!(float, 1.0, 4);
-    test_limit_is_reachable!(double, 1.0, 8);
-    test_limit_is_reachable!(bool, true, 1);
-    test_limit_is_reachable!(enumeration, proto::BadEnum::Long as i32, 10);
-}
+mod limit;
+mod overflow;
